Refacto to create the category and command files in the model entities
